pass half test point, basic point location

+++ Voronoi graph generation/site.h
@@ -5,5 +5,5 @@ public:
 	int id;
 	point p;
 	site() { id = 0; };
-	site(int x, int y, int num) : p(x,y), id(num){};
+	site(double x, double y, int num) : p(x,y), id(num){};
 };
